Add the token lifetime when checking it
Added the ability to pass a parameter containing the allowed token
lifetime in seconds to the token check function.
During check, the time elapsed since the token was created will be
compared with the time specified in this parameter. If the elapsed
time exceeds the specified parameter, the check returns False.
This parameter is optional and by default is equal to the value
specified in the PASSWORD_RESET_TIMEOUT variable