Add resending of registration confirmation email
The validity time of the confirmation email is set to 10 minutes.
After this time, it becomes invalid and the user account is deleted.
Added a button to resend the confirmation email. The button waiting time
is set to 5 minutes. A dynamic timer is implemented on the button, which
displays the remaining waiting time. Added the display of errors when
sending an email and the display of its successful sending.
You can resend the email once.