Add password reset function
A button "Forgot password," has been added to the login page,
clicking on which the user is redirected to the email input page.
After entering the email address and clicking on the button,
the user receives an email with a confirmation link.
By clicking on the link, the user must enter a new password twice.
In addition, the email sending function has been redesigned so that
it can be used in different ways. Now, when calling it, you need to
specify the subject and template of the message.