Add announcement addition when success validation
Added the function of creating a seller, if there is none.
Added the function of creating an announcement.
Added the function of creating announcement images.

All records in the database are created only if the fields
are successfully checked on the announcement creation page.

Also fixed an error related to the appearance of an error
in the location field if you select the communication
method and then the location.