Add user permissions
Added redirection of the user from the registration and login page
to the main page using the decorator, if he is already logged in.
Added redirection of the user from the page of adding an announcement
to the login page if he is not logged in.
Added error display if an unauthorized user tries to get the seller's
contact information or add an announcement to favorites.